requirements.txt
Jinja2==3.1.6
ldap3==2.9.1
mysql-connector-python==9.2.0
openiap==0.0.34
oracledb==2.5.1
plyer==2.1.0
psutil==6.0.0
psycopg2==2.9.9
psycopg2_binary==2.9.9
PyGetWindow==0.0.9
pymongo==4.15.0
pyodbc==5.2.0
python-dotenv==1.1.1
pywin32==308
Requests==2.32.5
sshtunnel==0.4.0
streamlit==1.42.2
keyboard==0.13.5
configparser==5.3.0
watchdog==3.0.0
streamlit_extras==0.5.0
